webpack + modules
