+++ src/common/index.js
@@ -1,0 +1,4 @@
+const colorList = [`black`, `yellow`, `blue`, `green`, `pink`];
+const daysList = [`mo`, `tu`, `we`, `th`, `fr`, `sa`, `su`];
+
+export {colorList, daysList};
