+++ src/components/card/control-bar-template/index.js
@@ -1,0 +1,12 @@
+function controlBarTemplate() {
+  return `
+  <div class="card__color-bar">
+    <svg class="card__color-bar-wave" width="100%" height="10">
+      <use xlink:href="#wave"></use>
+    </svg>
+  </div>
+  `;
+}
+
+export default controlBarTemplate;
+
