+++ src/components/card/status-btns-template/index.js
@@ -1,0 +1,10 @@
+function statusBtnsTemplate() {
+  return `
+  <div class="card__status-btns">
+    <button class="card__save" type="submit">save</button>
+    <button class="card__delete" type="button">delete</button>
+  </div>
+  `;
+}
+
+export default statusBtnsTemplate;
