+++ src/filter-list.js
@@ -1,0 +1,42 @@
+const filtersList = [
+  {
+    id: `filter__all`,
+    name: `ALL`,
+    count: 15,
+    checked: true
+  },
+  {
+    id: `filter__overdue`,
+    name: `OVERDUE`,
+    count: 0,
+    disabled: true
+  },
+  {
+    id: `filter__today`,
+    name: `TODAY`,
+    count: 0,
+    disabled: true
+  },
+  {
+    id: `filter__favorites`,
+    name: `FAVORITES`,
+    count: 7
+  },
+  {
+    id: `filter__repeating`,
+    name: `Repeating`,
+    count: 2
+  },
+  {
+    id: `filter__tags`,
+    name: `Tags`,
+    count: 6
+  },
+  {
+    id: `filter__archive`,
+    name: `ARCHIVE`,
+    count: 115
+  },
+];
+
+export default filtersList;
