+++ src/index.html
@@ -1922,5 +1922,5 @@ Here is a card with filled data</textarea
         <button class="load-more" type="button">load more</button>
       </section>
     </main>
-  <script type="text/javascript" src="bundle.js"></script></body>
+  </body>
 </html>
